LPS-115840 Move methods used to retrieve JournalArticle fieldValues to an implementation of InfoItemFieldValuesProvider
Extract field definitions into the new JournalArticleInfoItemFields interface to avoid duplication
